refactor(domain): move bounded context files to subfolder
