+++ src/core/domain/bounded-context/BoundedContextProjection.ts
