+++ src/core/usecases/queries/GetAllBoundedContexts/GetAllBoundedContextsQueryHandler.ts
@@ -1,5 +1,5 @@
-import { type BoundedContextCollection } from '../../../domain/BoundedContextCollection';
-import { type BoundedContextProjection } from '../../../domain/BoundedContextProjection';
+import { type BoundedContextCollection } from '../../../domain/bounded-context/BoundedContextCollection';
+import { type BoundedContextProjection } from '../../../domain/bounded-context/BoundedContextProjection';
 import { type GetAllBoundedContextsQuery } from './GetAllBoundedContextsQuery';
 
 export class GetAllBoundedContextsQueryHandler {
