+++ src/ui/components/Canvas/buildBoundedContextRowRenderData.ts
@@ -1,6 +1,6 @@
 import { type Node } from '@xyflow/react';
 import { type BoardProjection } from '../../../core/domain/BoardProjection';
-import { type BoundedContextProjection } from '../../../core/domain/BoundedContextProjection';
+import { type BoundedContextProjection } from '../../../core/domain/bounded-context/BoundedContextProjection';
 import { type SwimlaneColor } from '../../../core/domain/SwimlaneColor';
 import { type BoundedContextRowBackgroundNodeData } from './BoundedContextRowBackgroundNode';
 import { type FixedRowLabelEntry } from './FixedRowLabelColumn';
